[refactor] booking guard/header seat-selection 의존 경로 정리 #114
